@@ -5,6 +5,7 @@
   "bin": {
     "template": "dist/src/bin/index.js"
   },
+  "type": "module",
   "main": "dist/src/index.js",
   "types": "dist/src/index.d.ts",
   "scripts": {
@@ -15,8 +16,8 @@
     "compile:watch": "tsc --skipLibCheck --watch",
     "test": "ava",
     "test:watch": "ava --watch",
-    "uncompile": "node -r ts-node/register --inspect ./src/bin/index.ts",
-    "tnt": "node -r ts-node/register --inspect ./src/bin/index.ts",
+    "uncompile": "node  --loader ts-node/esm --inspect ./src/bin/index.ts",
+    "tnt": "node  --loader ts-node/esm --inspect ./src/bin/index.ts",
     "coverage": "nyc --reporter=html ava",
     "report": "nyc report",
     "report:html": "open coverage/index.html",
@@ -50,35 +51,35 @@
   ],
   "dependencies": {
     "@caporal/core": "^2.0.2",
+    "@tib/configload": "^0.3.0",
     "absolute": "0.0.1",
     "consolidate": "^0.14.5",
     "del": "^3.0.0",
     "fs-finder": "^1.8.1",
-    "handlebars": "^4.0.10",
-    "inquirer-helpers": "^1.2.1",
+    "handlebars": "^4.7.7",
+    "inquirer-helpers": "^1.3.5",
     "is-directory": "^0.3.1",
-    "is-url": "^1.2.2",
-    "isbinaryfile": "^3.0.2",
-    "load-json-file": "^3.0.0",
-    "lodash": "^4.17.4",
-    "metalsmith": "^2.3.0",
+    "is-url": "^1.2.4",
+    "isbinaryfile": "^3.0.3",
+    "jsox": "^1.2.113",
+    "lodash": "^4.17.21",
+    "metalsmith": "^2.5.0",
     "multimatch": "^2.1.0",
-    "p-series": "^1.0.0",
+    "p-series": "^1.1.0",
     "pify": "^3.0.0",
-    "specialist": "^0.3.0"
+    "specialist": "^0.3.0",
+    "ts-node": "^10.9.1",
+    "typescript": "^4.7.4"
   },
   "devDependencies": {
-    "@types/lodash": "^4.14.118",
+    "@types/lodash": "^4.14.182",
     "@types/metalsmith": "^2.3.1",
-    "@types/node": "^10.12.8",
+    "@types/node": "^10.17.60",
     "ava": "^0.22.0",
-    "ava-spec": "^1.1.0",
-    "esbuild": "^0.14.48",
+    "ava-spec": "^1.1.1",
+    "esbuild": "^0.14.49",
     "eslint": "^8.19.0",
-    "nyc": "^11.1.0",
-    "rimraf": "^2.6.0",
-    "ts-node": "^10.8.2",
-    "typescript": "^3.0.3",
-    "typescript-transform-export-interop": "^1.0.1"
+    "nyc": "^11.9.0",
+    "rimraf": "^2.7.1"
   }
 }
