@@ -1,18 +1,28 @@
 
 /* IMPORT */
 
-const caporal = require("@caporal/core").default
+
+import pkg from '@caporal/core';
+const { program } = pkg;
+
 import {updater} from 'specialist';
-import {name, version} from '../package.json';
 import Template from '.';
+import {autoLoadSync} from '@tib/configload';
 
 /* CLI */
 
-async function CLI () {
-  
-  updater ({ name, version });
+export async function CLI () {
 
-    caporal
+//   const myjson = await import('../package.json', {
+//    assert: { type: 'json' }
+//  });
+   const myjson = await autoLoadSync('package.json') as any
+   const name = myjson.name;
+   const version = myjson.version;
+   updater ({ name , version });
+
+
+   program
         /* WIZARD */
     .action ( () => Template.wizard () )
         /* CREATE */
@@ -20,7 +30,7 @@ async function CLI () {
     .argument ( '<template>', 'Template name' )
     .argument ( '[path2TempDir]', 'Path To Template Dir' )
     .option ( "-dd, --dontDelete", "Do Not delete files ,useful when upgrading templates")
-    .action ( ({ args, options }) => Template.create(args.template, args.path2TempDir, options.dontDelete))
+    .action ( ({ args, options }) => Template.create(args.template as string, args.path2TempDir as string, options.dontDelete as boolean))
     /* LIST */
     .command ( 'list', 'List installed templates' )
     .action ( () => Template.list () )
@@ -31,21 +41,21 @@ async function CLI () {
         .action(({logger, args }) =>
         {
             logger.info("intalling template from , %r!", args.repository)
-            Template.install(args.repository, args.template)
+            Template.install(args.repository as string, args.template as string)
         })
     /* UNINSTALL */
     .command ( 'uninstall', 'Uninstall one or all templates' )
     .argument ( '[template]', 'Template name' )
         .action(({ logger, args }) => {
             logger.info("unintalling template from , %r!", args.repository)
-            Template.uninstall(args.template)
+            Template.uninstall(args.template as string)
         })
     /* UPDATE */
     .command ( 'update', 'Update one or all templates' )
     .argument ( '[template]', 'Template name' )
         .action(({ logger, args }) => {
             logger.info("updateing template from , %r!", args.repository)
-            Template.update(args.template)
+            Template.update(args.template as string)
         })
     /* GENERATE */
     .command('generate', 'Generate files from a template')
@@ -54,12 +64,9 @@ async function CLI () {
     .argument('[path2TempDir]', 'Path to dir in wich template file are generated')
     .argument('[defaults]', 'Template default values')
     .argument('[files]', ' default values')
-    .action(({ args }) => Template.generate(args.template, args.path2TempDir, args.defaults, args.files))
-    
-  caporal.run(process.argv.slice(2))
+    .action(({ args }) => Template.generate(args.template as string, args.path2TempDir as string, args.defaults as string, args.files as string))
+
+    program.run(process.argv.slice(2))
 
 }
 
-/* EXPORT */
-
-export default CLI;
