@@ -3,32 +3,38 @@
 
 //import { default as prompt } from './prompt';
 //import { default as render } from './render';
-import { loadSchema } from './schema';
-import {templateLoader} from"./template"
+import { schemaManager  as schemaManager } from './schema';
+import { hookManager ,hooks} from"./hooks"
+import { Files ,} from 'metalsmith';
 import * as path from 'path';
-import { Files ,File } from 'metalsmith';
-import Metalsmith = require('metalsmith');
+import * as Metalsmith from 'metalsmith';
+import * as _ from 'lodash'
 
-
-export function generator(files:Files ,metalsmith:Metalsmith, next):void{
+export async function generator(files:Files ,ms:Metalsmith, next):Promise<void>{
     
-    const source = metalsmith.source(),
-    templatePath = path.dirname ( source )
+    const msPaths={
+    source : ms.source(),
+    template :path.dirname ( ms.source())
+   }
     //template = path.basename(schema);
-    const metadata= metalsmith.metadata () as any
-    const schema = loadSchema(templatePath)
+    //const metadata= metalsmith.metadata () as JSON
+    const schema = schemaManager.get(msPaths.template)
     
-   
-    files.file.forEach((file: File) => {
-    schema.isFileVailid(file.filepath,file.contents)
-        metadata.schema = schema.getSchema(file.filepath, file.contents);
-        const template = new templateLoader(templatePath)
-        if (template.hasHook()) { 
-            
+    _.forOwn ( files, (file, filepath) => {
+
+        if (schema.isFileVailid(filepath, file.contents)) {
+
+            //metadata.schema = schema.getSchema(filepath, file.contents);
+            const hook = new hookManager(msPaths.template)
+
+            hook.loadHooks().then( (hook:hooks) =>{
+                hook.StartPraser(file.filepath,schema)
+            })
+
         }
-        
-    
-    })
+
+
+    });
     //forEach(files => {
     //    schema.load(templatePath,file.filepath)
         
